requirements.txt
instagrapi==1.16.30
python-dotenv==0.21.0